generator/requirements.txt
google_api_python_client==2.116.0
google_auth_oauthlib==1.2.0
pandas==2.2.0
protobuf==4.25.2
